Student-achievement-system
This app is basically for student and for department. This is used to store the activity done by student so that whenever the student create resume so at that time, he/she may get data easily and also generate one-page kind of report of all activities done by student. When any grading system comes so that they can fetch data easily.